nerdctl run works
